fix: feature sanity check in rule based toolchains
This check ensures feature names are not duplicated. It does this by collecting all features. Including those from capabilities from tools from the `tool_map`.

If we now annotate a `gcc` tool with `@rules_cc//cc/toolchains/capabilities:supports_pic` and that tool is used for `c_compile_actions`, `cpp_compile_actions` and `link_actions`, the capability is collected multiple times and so is the feature.

If we extend the test here, we exclude the override check on the same feature.